Per #1886, port over the bugfixes from main_v10.0 to develop.

+++ met/internal_tests/libcode/vx_series_data/test_series_data.cc
@@ -46,7 +46,7 @@ int main(int argc, char *argv[]) {
     double delta = (max - min) / nbin;
     double width = 10;
 
-    vector<int> pdf;
+    vector<long long> pdf;
 
     init_pdf(nbin, pdf);
